requirements.txt
# ATM Transaction Pattern Analysis Project Requirements
# Core dependencies for professional analytics dashboard

# Database connectivity
mysql-connector-python==8.2.0

# Data manipulation and analysis
pandas==2.2.0
numpy==1.26.0

# Visualization
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.0

# Web framework
streamlit==1.34.0

# Additional utilities
python-dateutil==2.8.2
